requirements/test.txt
coverage==6.0.2
flake8==3.9.2
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
pytest-mock==3.6.1
pytest-deadfixtures==2.2.1
pytest-randomly==3.10.1
pytz==2021.3
django-fsm==2.7.1
django-model-utils==4.1.1
djangorestframework==3.12.4
pyyaml==5.4.1
uritemplate==4.1.0
flake8-bugbear==21.9.2
flake8-broken-line==0.3.0
flake8-commas==2.0.0
flake8-comprehensions==3.7.0
flake8-debugger==4.0.0
flake8-eradicate==1.1.0
flake8-isort==5.0.3
flake8-string-format==0.3.0
importlib-metadata>=4.0.0,<5.0.0
isort==5.10.1
